+++ CHANGES.md
@@ -16,6 +16,9 @@
 - [CHANGE] libwebrtc m93 で `__config_site` が必要になったため追加
     - zakuro からの移植
     - @melpon @voluntas
+- [CHANGE] libwebrtc m93 で api/video-track_source_proxy.h が pc/video_track_source_proxy.h に移動したのを修正
+    - zakuro からの移植
+    - @melpon @voluntas
 
 ## 2021.4.2