+++ CHANGES.md
@@ -13,6 +13,8 @@
 
 - [UPDATE] `libwebrtc` を `m93.4577.8.1` に上げる
     - @voluntas
+- [UPDATE] signaling mid 対応
+    - @melpon
 - [CHANGE] libwebrtc m93 で `__config_site` が必要になったため追加
     - zakuro からの移植
     - @melpon @voluntas